align date picker writeValue behavior with datetime picker (#9472)
when the date picker is supposed to write a value that is not pickable
then it will set an empty string. Contrary to this the date time picker
sets the value. Since the disable function should only be disabling
what the user can pick, programatically set values (e.g.: through the
form control) should be set and visible while validators handle what to
do with such a value

Co-authored-by: Inna Atanasova <39598672+InnaAtanasova@users.noreply.github.com>
Co-authored-by: deno <mladen.droshev@sap.com>
Co-authored-by: Mike O'Donnell <mikerodonnell89@users.noreply.github.com>