fix(platform): fix fdp-table popover alignment
closes [#11230](https://github.com/SAP/fundamental-ngx/issues/11230)

## Description
Ensure that the column header alignment is preserved when sorting columns in the fdp-table component. Previously, the header would jump to the start regardless of the initial alignment when sorting was activated. The header styles have been adjusted to maintain the specified alignment.

Additionally, the popover used for sorting is now aligned according to the column alignment, rather than always defaulting to the left.